Added testing_api.py to project planning folder
Can be used as a framework for how we are going to query the api.
When we are ready, we should be able to follow this framework in the web
app views.  Note that we will all need to install requests in our
virtual environments.  Use the command
conda install requests