Operator Query Tool Integration Testing
I have created a new directory for testing, where I have documented the
testing I performed on the operator query tool.  This was done using the
behaviour driven development model.  The tests were carried out using
selenium and are contained in the tests.py file for the operator query
tool.