Upgraded security of operator api
I have added the use of tokens to the operator api.  Now if the api will
not allow the user to access the data if they do not have a verified
token.  All users logged into the query tool will have a token.  They
have already been authenticated so the requests are more secure now.