Limited number of requests on operator API
Previously the API was vulnerable to DoS attacks.  Now I have upgraded
the security by ensuring that there is a limit to the amount of requests
allowed per day.  I have set the limit to 200 just as a random number,
but more research can be done into discovering how much we can push our
API.