Customer Query Tool Being Prepped to Query other operators
I have been adapting the customer query tool to add the option of
linking to other customer accounts that you may have.  This will allow
us to then implement the feature of the customer account querying other
accounts.  Currently, the customer can select the account they want to
connect to.  An email is sent to them with a code which they can then
use to verify the account.  Then the two accounts are linked.