Working on two accounts communicating
I have set up the customer account for PSD ferries so that it can link
to other customer accounts and retrieve data from them.  It is also
capable of being queried itself.  The next steps will be to create a new
account, and also two population scripts so we can simulate the flow of
data between them.