Added Second Customer Account For Communication
I have added the second customer account.  Both the new bus account and
the ferry account have appropriate population scripts.  Theoretically,
once we set them up and running on python anywhere, they will be able to
communicate with each other.  Before this can be done, we need to add
both of these accounts to the operator database so they can be found.